pyproject.toml
[project]
name = "custommcp"
version = "0.1.0"
description = "custommcp using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.121.0,<1.0.0",
    "fastmcp>=2.5.2",
]

[project.scripts]
custommcp = "custommcp.main:run"
run_crew = "custommcp.main:run"
train = "custommcp.main:train"
replay = "custommcp.main:replay"
test = "custommcp.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
